refactor: move cluster metadata parsing code to misc dir
